unconvert

+++ internal/classifier/cel_lists.go
@@ -303,8 +303,8 @@ func indexOf(list ref.Val, item ref.Val) ref.Val {
 
 	sz := lister.Size().(types.Int)
 	for i := types.Int(0); i < sz; i++ {
-		if lister.Get(types.Int(i)).Equal(item) == types.True {
-			return types.Int(i)
+		if lister.Get(i).Equal(item) == types.True {
+			return i
 		}
 	}
 
@@ -320,8 +320,8 @@ func lastIndexOf(list ref.Val, item ref.Val) ref.Val {
 	sz := lister.Size().(types.Int)
 
 	for i := sz - 1; i >= 0; i-- {
-		if lister.Get(types.Int(i)).Equal(item) == types.True {
-			return types.Int(i)
+		if lister.Get(i).Equal(item) == types.True {
+			return i
 		}
 	}
